+++ CHANGELOG.md
@@ -1,1 +1,3 @@
-* Update toc file.
+* Expeditions list sort by name.
+* Fixed summon on double right click in journal.
+* Support 8.2 features.